Implement a USB connection for the mkII ICE.
As the main code wants to select() for events, and libusb doesn't
provide us with a simple file descriptor we could select on, the USB
communication uses a separately forked-off USB daemon that handles the
libusb-based USB IO, and talks through a bidirectional pipe
(aka. socketpair()) with the main AVaRICE code.